+++ wkdan_Java/src/Main/java/BOJ_01/p1001.java
@@ -1,15 +1,19 @@
-package BJ_01;
+package BOJ_01;
 
 import java.io.*;
+import java.util.StringTokenizer;
 
-public class p10926 {
+
+public class p1001 {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
 
-        String id = br.readLine();
+        StringTokenizer st = new StringTokenizer(br.readLine());
+        int A = Integer.parseInt(st.nextToken());
+        int B = Integer.parseInt(st.nextToken());
 
-        bw.write(id+"??!"+"\n");
+        bw.write(A-B+"\n");
         bw.flush();
         bw.close();
