+++ wkdan_Java/src/Main/java/BOJ_01/p1008.java
@@ -1,25 +1,20 @@
-package BJ_01;
-
+package BOJ_01;
 import java.io.*;
 import java.util.StringTokenizer;
 
-public class p10430 {
+
+public class p1008 {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
 
         StringTokenizer st = new StringTokenizer(br.readLine());
+        double A = Double.parseDouble(st.nextToken());
+        double B = Double.parseDouble(st.nextToken());
 
-        int A = Integer.parseInt(st.nextToken());
-        int B = Integer.parseInt(st.nextToken());
-        int C = Integer.parseInt(st.nextToken());
-
-        bw.write((A+B)%C+"\n");
-        bw.write(((A%C)+(B%C))%C+"\n");
-        bw.write(((A*B)%C)+"\n");
-        bw.write(((A%C)*(B%C))%C+"\n");
-
+        bw.write(A/B+"\n");
         bw.flush();
         bw.close();
+
     }
 }
