+++ wkdan_Java/src/Main/java/BOJ_01/p10430.java
@@ -1,20 +1,25 @@
-package BJ_01;
+package BOJ_01;
+
 import java.io.*;
 import java.util.StringTokenizer;
 
-
-public class p1008 {
+public class p10430 {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
 
         StringTokenizer st = new StringTokenizer(br.readLine());
-        double A = Double.parseDouble(st.nextToken());
-        double B = Double.parseDouble(st.nextToken());
 
-        bw.write(A/B+"\n");
+        int A = Integer.parseInt(st.nextToken());
+        int B = Integer.parseInt(st.nextToken());
+        int C = Integer.parseInt(st.nextToken());
+
+        bw.write((A+B)%C+"\n");
+        bw.write(((A%C)+(B%C))%C+"\n");
+        bw.write(((A*B)%C)+"\n");
+        bw.write(((A%C)*(B%C))%C+"\n");
+
         bw.flush();
         bw.close();
-
     }
 }
