+++ wkdan_Java/src/Main/java/BOJ_01/p11382.java
@@ -1,22 +1,20 @@
-package BJ_01;
+package BOJ_01;
+
 import java.io.*;
 import java.util.StringTokenizer;
 
-
-public class p10869 {
+public class p11382 {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
 
         StringTokenizer st = new StringTokenizer(br.readLine());
-        int A = Integer.parseInt(st.nextToken());
-        int B = Integer.parseInt(st.nextToken());
 
-        bw.write(A+B+"\n");
-        bw.write(A-B+"\n");
-        bw.write(A*B+"\n");
-        bw.write(A/B+"\n");
-        bw.write(A%B+"\n");
+        long A = Long.parseLong(st.nextToken());
+        long B = Long.parseLong(st.nextToken());
+        long C = Long.parseLong(st.nextToken());
+
+        bw.write(A+B+C+"\n");
         bw.flush();
         bw.close();
     }
