+++ wkdan_Java/src/Main/java/BOJ_01/p18108.java
@@ -1,21 +1,17 @@
-package BJ_01;
+package BOJ_01;
 
 import java.io.*;
-import java.util.StringTokenizer;
 
-
-public class p1001 {
+public class p18108 {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
 
-        StringTokenizer st = new StringTokenizer(br.readLine());
-        int A = Integer.parseInt(st.nextToken());
-        int B = Integer.parseInt(st.nextToken());
+        int year = Integer.parseInt(br.readLine());
+        int answer = year-543;
 
-        bw.write(A-B+"\n");
+        bw.write(answer+"\n");
         bw.flush();
         bw.close();
-
     }
 }
